+++ super_editor/lib/super_editor.dart
@@ -7,7 +7,6 @@ export 'package:super_text_layout/src/caret_layer.dart';
 export 'src/core/document.dart';
 export 'src/core/document_composer.dart';
 export 'src/core/document_debug_paint.dart';
-export 'src/core/document_editor.dart';
 export 'src/core/document_interaction.dart';
 export 'src/core/document_layout.dart';
 export 'src/core/document_selection.dart';
@@ -57,6 +56,7 @@ export 'src/infrastructure/focus.dart';
 export 'src/infrastructure/ime_input_owner.dart';
 export 'src/infrastructure/keyboard.dart';
 export 'src/infrastructure/multi_tap_gesture.dart';
+export 'src/infrastructure/pausable_value_notifier.dart';
 export 'src/infrastructure/platforms/android/android_document_controls.dart';
 export 'src/infrastructure/platforms/ios/ios_document_controls.dart';
 export 'src/infrastructure/platforms/mobile_documents.dart';
